+++ src/templates/domain/models/UpdateTemplateMetadataInfo.ts
@@ -1,0 +1,7 @@
+import { TemplateFieldInfo, TemplateInstructionInfo } from './TemplateInfo'
+
+export interface UpdateTemplateMetadataInfo {
+  name?: string
+  fields?: TemplateFieldInfo[]
+  instructions?: TemplateInstructionInfo[]
+}